requirements.txt
streamlit
google-generativeai
python-dotenv
nest_asyncio
PyPDF2
sendgrid
supabase
pytz
extra-streamlit-components
elevenlabs
miniaudio
imageio-ffmpeg